Regenerate tf files from example files
When an example file changes, we want to regenerate the corresponding
non-example file in order to easily pick  up the changes without having
to manually inspect the example file for changes and copy/paste
additions and modifications. This change will overwrite existing files,
except in the case of `cumulus-tf/terraform.tfvars` for any deployment,
and except for UAT (in which case none is overwritten).

In the future, this process will be simplified to better support
environement differences by simply supplying all terraform variables via
environment variables.