Export all vars from the `.env` file
In order for all AWS and Terraform commands to see all the necessary
environment variables, we must export them all. I'm not particularly
happy with how we're using `.env` and `bin/env.sh`, so we may need to
revisit how we handle env vars, ideally in a bit simpler and less
brittle (?) way.