Exclude `_?B_Analytic` suffix from granule ID
Since the Planet metadata json files do _not_ include such a suffix, we
must exclude it from the granule ID, otherwise they will not be picked
up during discovery, and thus won't be ingested with the rest of the
granule files.

However, this affects _only_ the granule IDs, which equate to the CMR's
native IDs, not to the granule URs found inside of the CMR metadata
files, and can be completely different.