Auto-create AWS service-linked role for Elasticsearch
Since Amazon Elasticsearch Service does not use a VPC Endpoint, to use
ES within a VPC, we must create a service-linked role for it.  This must
be created only once per AWS account, so this Terraspace hook
automatically creates it, if it does not already exist.  This eliminates
an otherwise manual Cumulus setup step that is easy to miss, and results
in a deployment error for the data-persistence module, if missed.

In addition:

- Remove obsolete VS Code setting

  With the introduction of Terraspace in this repo, the Terraform module
  directories have moved, so this VS Code setting is obsolete.  Further,
  it appears that we do not have to update this list of root modules to
  reflect their new locations, but that we can remove the list completely
  without negatively impacting automatic formatting of Terraform files in
  VS Code (perhaps the Terraform Language Server is now smarter).

- Minor reorg and commenting of Dockerfile

  Added `bundle clean --force` command to cleanup unused Ruby
  gem files, which were causing dependency problems when attempting to run
  Terraspace commands without prefixing them with `bundle exec`.  Using
  `bundle exec` in CI is fine, but becomes tedious when running locally,
  so this addresses this annoyance.