Fix statuses for executions and granules
The output from the DiscoverAndQueueGranules and IngestAndPublishGranule
workflows were wrapped in an array due to the use of Parallel tasks.
This was preventing Cumulus from updating the statuses of successful
executions as well as the statuses of granules successfully ingested.
Executions were left as "running" and granules were left as "queued."
This fix "unwraps" the output from the Parallel tasks so that Cumulus
receives the expected output format.