Move discovery and queueing granules steps to ECS tasks
The discovery and queuing steps of the DiscoverAndQueueGranules Step
Function were not able to handle process of moderately-sized inputs
without crashing or timing out, so they were moved to ECS tasks.

This helped, but even the default EC2 instance size (t3.medium) was
still undersized, so bumping to t3.large and dialing up the CPU and
memory reservation settings for the tasks finally did the trick.

Also, set preferredQueueBatchSize to 5 (from default of 1) to add
some parallelization to the queuing step, without the need to increase
the lambda function concurrency, which helped avoid some contention.

Fixes #30