Fix build/deploy failures; add convenience scripts
When running things in a specific order (or _not_ in a specific order),
it was possible to see the following error while attempting to deploy
Cumulus using the `make all-up` or `make up-cumulus` commands:

```plain
Error: failed to execute "/bin/bash": bash: cannot set terminal process
group (1): Inappropriate ioctl for device
bash: no job control in this shell
$ yarn -s clean && yarn -s test && yarn install --production \
--modules-folder ${PWD}/${npm_package_config_lambda_archive_dir}/node_modules
```

Unfortunately, that message is incredibly unhelpful, but it turned out
to be caused by Terraform's attempt to create zip a directory when
packaging code for a Lambda function, where the directory contains at
least 1 symlink. This is a long-standing bug in Terraform, but easily
corrected for our situation.

The cause was the installation of node dependencies with symlinks
to their executable scripts. Since we don't use the scripts, supplying
the `--no-bin-links` option to `yarn install` avoids installing them.

In addition, another cause of build failures was due to the way in which
TypeScript performs incremental compilation. Long-story short, disabling
incremental compilation, and also correcting how build artifacts are
cleaned up (via the `yarn clean` command) addresses this problem.

Deploying the entire Cumulus stack can take over an hour, which leads to
the following error when running `make all-up`:

```plain
bundler: failed to load command: terraspace (/opt/terraspace/embedded/bin/terraspace)
/opt/terraspace/embedded/lib/ruby/gems/2.7.0/gems/terraspace-0.6.17/lib/terraspace/shell.rb:91:in `eof': execution expired (Timeout::Error)
        from /opt/terraspace/embedded/lib/ruby/gems/2.7.0/gems/terraspace-0.6.17/lib/terraspace/shell.rb:91:in `block in all_eof?'
        from /opt/terraspace/embedded/lib/ruby/gems/2.7.0/gems/terraspace-0.6.17/lib/terraspace/shell.rb:91:in `each'
```

It was not clear that this was due to Terraspace having a timeout, but
that's exactly the issue. By default, Terraspace times out after 1 hour,
and abruptly kills the deployment process. This is extremely problematic
when Terrform is in the midst of deployment with a lock on the state
file, as it does not gracefully allow Terraform to finish its work and
release its lock, thus making subsequent attempts to finish deployment
an absolute nightmare because Terraform is utterly out of touch with
reality at that point.

This not only bumps up the Terraspace timeout (`TS_BUFFER_TIMEOUT`) to
4 hours (14400 seconds), but also addresses the following: (1) upgrades
all out-of-date Ruby gems (in `Gemfile.lock`), (2) pins the Docker
source image, and (3) removes some bloat from the Docker image.

Added `bin/nuke.sh` to easily "nuke" a deployment, along with docs on
how to destroy managed and unmanaged resources.

During deployment, if Terraform is unable to gracefully complete its
work and release its lock on its state file, it can be extremely painful
and time-consuming to clean up the mess. Often, after force-releasing
the lock, subsequent attempts to resume deployment produce many
"duplicate resource" errors. The "Terraform Doctor" script makes it a
snap to fix most (if not all) such errors.

- Add missing Docker mounts to dev container
- Remove commented-out code
- Add missing yarn script
- Add missing CUMULUS_PREFIX var in `.env.example`