Remove UAT-specific prefix
In an attempt to correct the problems we're having recreating
subscriptions between our SNS topics and the Cloud Metrics SQS queues,
we're recreating the UAT stack with a different prefix. Specifically,
we're dropping the UAT-specific prefix and defaulting to the pattern
used by all other stacks (i.e., `cumulus-${TS_ENV}`).