Fix timeout, idempotency issues with "nuke" script
The nuke script will abruptly halt if a Terraspace command takes more
than an hour (by default) to complete, which causes problems. Bumped the
`TS_BUFFER_TIMEOUT` value to 14400 seconds (4 hours) to avoid this.

Further, added some logic to allow rerunning the script in the event of
failure so that it can pick up where it left off.