Fix "create test data" script for UAT variation
Although the prefix for UAT has changed for all resources managed by
Terraform, we have _not_ changed the bucket names so that we don't have
to submit new NAMS requests to setup things already in place, such as
CloudFront URLs. Therefore, we need to account for this difference in
UAT within this script.