Fix error handling in Cumulus CLI
Error-handling was, well, erroneous itself, which became evident when
attempting to "upsert" a provider in an empty deployment. An error was
produced during the "update" attempt (as expected since there's nothing
to update in an empty DB), but the error was not properly propogated.
This simply caused the upsert command to fail due to the update failure,
rather than subsequently making an insert attempt.

Also, added a "version" command to allow us to see what version of the
Cumulus API we're hitting.