Pin tfenv version to 2.2.3
Build started to fail because new commits to tfenv repo includes new log
messages that are written to stdout, which screws up Terraspace, which
parses JSON output written to stdout from Terraform. With the tfenv log
messages added to stdout, the output on stdout is no longer valid JSON,
and even if it were, it's not part of the Terraform output that Terraspace
expects.

Issue #67