Make granule queue batch size configurable
For granules with large data files, the time to compute a data file's
checksum can become significant, thus impacting the time for the
AddMissingFileChecksums lambda function. When granules are batched, this
can cause the lambda function to timeout (possibly even beyond the max
limit of 15 minutes), so we cannot assume that it can always handle a
fixed batch size of 5, as previously hard-coded.

This change allows configuration of the batch size on a per collection
basis, with a default size of 1, if not explicitly congifured.