Read granule file checksums from cmr.json files
Avoid potentially expensive checksum computation by simply populating
granle file checksums from the cmr.json file that is part of each
granule. If a checksum for a particular file is not found, a message is
logged, but no error is thrown, as any missing checksums will simply be
computed in the subsequent step of the workflow.