Add support for WorldView 4 (#80)
* Add WorldView4 resources

* Make granule queue batch size configurable

For granules with large data files, the time to compute a data file's
checksum can become significant, thus impacting the time for the
AddMissingFileChecksums lambda function. When granules are batched, this
can cause the lambda function to timeout (possibly even beyond the max
limit of 15 minutes), so we cannot assume that it can always handle a
fixed batch size of 5, as previously hard-coded.

This change allows configuration of the batch size on a per collection
basis, with a default size of 1, if not explicitly configured.

* Simplify granule discovery

Refactored granule discovery to provide simpler logic. Reverted back to
using the "out of box" DiscoverGranules lambda function, as our
replacement function was somewhat convoluted relative to the amount of
custom logic, which only adds prefixes to granule IDs.

The logic to prefix granule IDs has been greatly simplified and moved to
a separate lambda function, placed immediately after the DiscoverGranules
lambda function within the discover/queue granules workflow.

* Read granule file checksums from cmr.json files

Avoid potentially expensive checksum computation by simply populating
granle file checksums from the cmr.json file that is part of each
granule. If a checksum for a particular file is not found, a message is
logged, but no error is thrown, as any missing checksums will simply be
computed in the subsequent step of the workflow.

* In addition:
  * Replace dayjs lib with date-fns and duration-fns
  * Upgrade ava to 4.x
  * Add markdown extensions to dev container
  * Improve failure output from prettier
  * Make prefixing granule IDs configurable, per collection
  * Remove obsolete/unused code
  * Address "got" dependency vulnerability