Fix false-positive terraspace plan
The `terraspace plan` command exits with status code 0 even when the
wrapped `terraform plan` command fails with a non-zero status code.
Unfortunately, this means that the `terraspace plan` command _never_
fails, even when it _should_, thus allowing CI for UAT to succeed
when it should _fail_.

This fixes the "got" dependency issue that caused unit tests to fail,
subsequently causing `terraform plan` to fail, but also upgrades
Terraspace (and it's ruby gems) to the latest version, in which the
`terraspace plan` command is fixed to properly propagate the exit
status of the wrapped `terraform plan` command.