Fix terraspace plan failure
Ensure the version of Node specified in the `.nvmrc` file is installed
and used for deployment within GitHub Actions.

Also, adjust DateFormat test to avoid dependence on error message
wording from core Node error, which changed in v16. It is because of
this wording change, along with the upgrade to Node v16 in the GitHub
runner that the build started to fail.