Make launchpad.pfx a binary secret
Rather than storing launchpad.pfx as a file in some "common" bucket so
it can be copied to stack-specific buckets, we now store it as a binary
secret. This allows far more convenient usage, as secret names are not
required to be universally unique, whereas bucket names are. Thus, we
can use the same secret name, regardless of AWS account.

Copying launchpad.pfx to a stack-specific bucket (where Cumulus expects
to find it) now involves decoding the secret to a local file, then
uploading the file to the stack-specific S3 location, making this
operation completely account-independent (assuming the secret has been
created in advance).

Further, to make it easy to transfer the secret between accounts, the
transfer script automates this by requiring only the src and dest AWS
profiles as arguments.