Do not overwrite existing test fixture files
To allow real thumbnails to be used as test fixture files, and also to
accomodate thumbnail files that are not in the list of DataGranule files,
do not generate dummy files in place of existing fixture files, and do
not delete files that do not appear in the DataGranule file list.

Further, filter out `.DS_Store` files from the logic that locates the
directories containing granule files, as these cause the script to throw
errors regarding missing cmr.json files.